+++ CHANGELOG.md
@@ -14,9 +14,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Use bit-banding for Peripheral enable/reset.
   Don't require APBs in initializers.
 - Rename `gpio::Edge::{RISING, FALLING, RISING_FALLING}` to `Rising`, `Falling`, `RisingFalling`, respectively
+- Added the ability to specify the word size (8 or 9 bits) for `Serial` (USART). When using parity, the parity bit is included in the number of bits of the word.
 
 ### Added
 
+- `blocking::serial::Write` for `Tx` and `Serial`. `core::fmt::Write` for `Serial`
 - `rcc::Config` with prescalers for direct setting of clocks without calculating
 - `From<Bps>` for `serial::Config`
 - `From<Into<Hertz>>` for `i2c::Mode`
@@ -45,6 +47,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Fix SPI3 alternate function remapping
 - Do not enable UART DMA flags unconditionally
 - Fix flash erase verification always failing
+- Fix invalid 8-bit access to USART registers.
 
 ### Changed